@@ -7,26 +7,27 @@
     "develop": "strapi develop",
     "start": "strapi start",
     "build": "strapi build",
-    "strapi": "strapi",
-    "lint": "eslint --ext .js,.ts . --fix"
-  },
-  "devDependencies": {
-    "babel-eslint": "^10.1.0"
+    "strapi": "strapi"
   },
   "dependencies": {
-    "@strapi/plugin-i18n": "4.0.2",
-    "@strapi/plugin-users-permissions": "4.0.2",
-    "@strapi/strapi": "4.0.2",
+    "@strapi/plugin-i18n": "4.5.2",
+    "@strapi/plugin-users-permissions": "4.5.2",
+    "@strapi/strapi": "4.5.2",
+    "strapi-plugin-multi-select": "1.1.0",
+    "@strapi/design-system": "1.3.1",
+    "@strapi/icons": "1.3.1",
+    "@floating-ui/react-dom": "1.0.1",
+    "@fortawesome/react-fontawesome": "0.2.0",
     "pg": "8.6.0"
   },
   "author": {
     "name": "A Strapi developer"
   },
   "strapi": {
-    "uuid": ""
+    "uuid": "4c18d5b8-edd2-4222-a771-5a931317b075"
   },
   "engines": {
-    "node": ">=12.x.x <=16.x.x",
+    "node": ">=18.0.0",
     "npm": ">=6.0.0"
   },
   "license": "MIT"
