@@ -1,6 +1,6 @@
 {
   "name": "frontend",
-  "version": "0.0.0",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -9,17 +9,30 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "12.0.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "ui": "*"
+    "@nextui-org/react": "1.0.0-beta.10",
+    "@types/styled-components": "^5.1.26",
+    "@types/styled-react-modal": "^1.2.2",
+    "axios": "^1.1.3",
+    "mongoose": "^6.7.1",
+    "next": "13.0.2",
+    "next-auth": "^4.15.1",
+    "phosphor-react": "^1.4.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-fontawesome": "^1.7.1",
+    "react-modal": "^3.16.1",
+    "styled-components": "^5.3.6",
+    "styled-react-modal": "^3.0.1"
   },
   "devDependencies": {
-    "@types/react": "17.0.37",
-    "config": "*",
-    "eslint": "7.32.0",
-    "next-transpile-modules": "9.0.0",
-    "tsconfig": "*",
-    "typescript": "^4.5.3"
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "babel-plugin-styled-components": "^2.0.7",
+    "eslint": "8.26.0",
+    "eslint-config-next": "13.0.2",
+    "typescript": "^4.8.4"
+  },
+  "compiler": {
+    "styledComponents": true
   }
 }
