@@ -1,4 +1,4 @@
-"use strict";
+'use strict';
 
 /* eslint-disable no-unused-vars */
 module.exports = (config, webpack) => {
