@@ -1,5 +1,16 @@
-const withTM = require("next-transpile-modules")(["ui"]);
-
-module.exports = withTM({
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   reactStrictMode: true,
-});
+  swcMinify: true,
+
+  images: {
+    domains: ["avatars.githubusercontent.com"],
+  },
+
+  compiler: {
+    // see https://styled-components.com/docs/tooling#babel-plugin for more info on the options.
+    styledComponents: true,
+  },
+};
+
+module.exports = nextConfig;
